+++ mapiproxy/servers/default/emsmdb/oxomsg.c
@@ -136,8 +136,8 @@ static void oxomsg_mapistore_handle_message_relocation(struct emsmdbp_context *e
 }
 
 /**
-   \details EcDoRpc SubmitMessage (0x32) Rop. This operation marks a message
-   as being ready to send (subject to some flags).
+   \details EcDoRpc SubmitMessage (0x32) Rop. This operation sends a message
+   to its designated recipients. It is usually performed in online mode.
 
    \param mem_ctx pointer to the memory context
    \param emsmdbp_ctx pointer to the emsmdb provider context
@@ -163,8 +163,6 @@ _PUBLIC_ enum MAPISTATUS EcDoRpc_RopSubmitMessage(TALLOC_CTX *mem_ctx,
 	void			*private_data;
 	bool			mapistore = false;
 	struct emsmdbp_object	*object;
-	char			*owner;
-	uint64_t		messageID;
 	uint32_t		contextID;
 	uint8_t			flags;
 
@@ -224,12 +222,14 @@ _PUBLIC_ enum MAPISTATUS EcDoRpc_RopSubmitMessage(TALLOC_CTX *mem_ctx,
 			}
 		}
 
-		messageID = object->object.message->messageID;
+		retval = emsmdbp_object_attach_sharing_metadata_XML_file(emsmdbp_ctx, object);
+		if (retval != MAPI_E_SUCCESS) {
+			OC_DEBUG(1, "Failing to create sharing metadata for a sharing object: %s\n", mapi_get_errstr(retval));
+		}
+
 		contextID = emsmdbp_get_contextID(object);
 		flags = mapi_req->u.mapi_SubmitMessage.SubmitFlags;
-		owner = emsmdbp_get_owner(object);
-
-		ret = mapistore_message_submit(emsmdbp_ctx->mstore_ctx, emsmdbp_get_contextID(object),
+		ret = mapistore_message_submit(emsmdbp_ctx->mstore_ctx, contextID,
 					       object->backend_object, flags);
 		if (ret != MAPISTORE_SUCCESS) {
 			OC_DEBUG(1, "Failing to submit the message: %s", mapistore_errstr(ret));
@@ -238,7 +238,6 @@ _PUBLIC_ enum MAPISTATUS EcDoRpc_RopSubmitMessage(TALLOC_CTX *mem_ctx,
 		}
 
 		oxomsg_mapistore_handle_message_relocation(emsmdbp_ctx, object);
-		mapistore_indexing_record_add_mid(emsmdbp_ctx->mstore_ctx, contextID, owner, messageID);
 		break;
 	}
 
@@ -373,7 +372,9 @@ _PUBLIC_ enum MAPISTATUS EcDoRpc_RopGetAddressTypes(TALLOC_CTX *mem_ctx,
 }
 
 /**
-   \details EcDoRpc TransportSend (0x4a) Rop. This operation sends a message.
+   \details EcDoRpc TransportSend (0x4a) Rop. This operation requests
+   to the server to send a message to recipients. It is usually
+   performed in cached mode.
 
    \param mem_ctx pointer to the memory context
    \param emsmdbp_ctx pointer to the emsmdb provider context
@@ -438,7 +439,7 @@ _PUBLIC_ enum MAPISTATUS EcDoRpc_RopTransportSend(TALLOC_CTX *mem_ctx,
 	case true:
 		retval = emsmdbp_object_attach_sharing_metadata_XML_file(emsmdbp_ctx, object);
 		if (retval != MAPI_E_SUCCESS) {
-			OC_DEBUG(0, "Failing to create sharing metadata for a sharing object: %s\n", mapi_get_errstr(retval));
+			OC_DEBUG(1, "Failing to create sharing metadata for a sharing object: %s\n", mapi_get_errstr(retval));
 		}
 
 		ret = mapistore_message_submit(emsmdbp_ctx->mstore_ctx, emsmdbp_get_contextID(object), object->backend_object, 0);
@@ -449,7 +450,6 @@ _PUBLIC_ enum MAPISTATUS EcDoRpc_RopTransportSend(TALLOC_CTX *mem_ctx,
 		}
 
 		oxomsg_mapistore_handle_message_relocation(emsmdbp_ctx, object);
-		/* mapistore_indexing_record_add_mid(emsmdbp_ctx->mstore_ctx, contextID, messageID); */
 		break;
 	}